Move JsonInheritanceAttribute without a namespace change
